create helper module for the JSON output of report
