+++ _scripts/_outputs/format-JSON.js
@@ -1,0 +1,12 @@
+function formatJSON(imageStats, results) {
+    var outputJSON = {};
+    // Ensure that the imageStats Object is not empty
+   if (!(Object.keys(imageStats).length === 0 && imageStats.constructor === Object)) {
+        outputJSON.images = imageStats;
+    }
+    outputJSON.results = results;
+    outputJSON = JSON.stringify(outputJSON, null, 2);
+    return outputJSON;
+}
+
+module.exports = formatJSON;
